break(@kikd/dependency-list): Return split deps between local/external.
Before we returned just an array of all dependencies, but downstream
packages had to figure out what was external and what was local. But now
we do it in the upstream and make it available. If all deps are needed
then they can be combined.

refactor(@kikd/template-dependencies): Simplify with the new deps obj.

refactor(@kikd/recommend-version): Use the local portion of new deps.