refactor(@kikd/build-rollup-configs): Adjust externals with new deps.
Before this change we included EVERY dependency in the root
`package.json` as an external dependency. Now we use the new deps object
to determine everything not in the project as a package as an external
dependency. This has the side effect of cleaning up the
[**Rollup**](https://rollupjs.org/) output when building with the weird
imports that include pathing like `loud-rejection/register`.

Closes #229