feat(@kikd/post-build): Dependencies in packages are now unique.
The resulting `package.json` file for each package used to include all
dependencies for the entire monorepo, but now it only includes the
dependencies for that particular package. This makes the end result much
more accurate. Since all local projects are rolled up, local projects
aren't in that list and their external dependencies are instead added.