feat(@kape/run): Remove need for array of arrays.
Closes #190

test(@kikd/check-package): Rename it to snapshot and remove array.

test(@kikd/dependency-detective): Rename it to snapshot and remove array.

test(@kikd/dependency-list): Rename it to snapshot and remove array.

test(@kikd/find-packages): Rename it to snapshot and remove array.

test(@kikd/publish-check): Rename it to snapshot and remove array.

test(@kikd/publishing): Rename it to snapshot and remove array.

test(@semcom/commit-parser): Rename it to snapshot and remove array.

test(@semcom/latest-release-tag): Rename it to snapshot and remove array.

test(@semcom/parse-tags): Rename it to snapshot and remove array.

test(@semcom/recommend-version): Rename it to snapshot and remove array.